Rename typing
