--- src/cladetime/typing.py
@@ -1,9 +1,0 @@
-"""Type aliases for this package."""
-
-from enum import StrEnum
-
-
-class StateFormat(StrEnum):
-    ABBR = "abbr"
-    NAME = "name"
-    FIPS = "fips"
